requirements.txt
langchain
langchain-groq
python-dotenv
chromadb
langchain-community
beautifulsoup4
langchain_core
pandas
uuid
streamlit 